Candy Spawner
Implemented a candy spawner that checks and replaces the empty slots with random candies. The method HandlepostMoveCheck in PostMatchDrop checks for matches, destroys them, drops the remaining candies, checks for matches, and repeats this loop until there are no matches. Then the spawner fills the empty slots and the first loop restarts.